requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
SQLAlchemy==2.0.32
loguru==0.7.2
pytest==8.3.2
pytest-asyncio==0.23.8
